fix: was not invalidating tags after import
